Add CI/CD workflow for app distribution
