+++ android/app/google-services.json
@@ -1,0 +1,29 @@
+{
+  "project_info": {
+    "project_number": "37031961174",
+    "project_id": "go-egypt-9e91b",
+    "storage_bucket": "go-egypt-9e91b.firebasestorage.app"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:37031961174:android:322859a270216a15ab51f7",
+        "android_client_info": {
+          "package_name": "com.example.go_egypt"
+        }
+      },
+      "oauth_client": [],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBZagDjK7DWqKU-eAbQBVN73CemGRv_7sE"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": []
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
