+++ android/fastlane/Appfile
@@ -1,0 +1,2 @@
+json_key_file("") # Path to the json secret file - Follow https://docs.fastlane.tools/actions/supply/#setup to get one
+package_name("com.example.go_egypt") # e.g. com.krausefx.app
