+++ firebase.json
@@ -1,0 +1,1 @@
+{"flutter":{"platforms":{"android":{"default":{"projectId":"go-egypt-9e91b","appId":"1:37031961174:android:322859a270216a15ab51f7","fileOutput":"android/app/google-services.json"}},"dart":{"lib/firebase_options.dart":{"projectId":"go-egypt-9e91b","configurations":{"android":"1:37031961174:android:322859a270216a15ab51f7","ios":"1:37031961174:ios:5db9549fadf01f3cab51f7"}}}}}}
