+++ lib/firebase_options.dart
@@ -1,0 +1,70 @@
+// File generated by FlutterFire CLI.
+// ignore_for_file: type=lint
+import 'dart:io';
+
+import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
+import 'package:flutter/foundation.dart'
+    show defaultTargetPlatform, kIsWeb, TargetPlatform;
+
+/// Default [FirebaseOptions] for use with your Firebase apps.
+///
+/// Example:
+/// ```dart
+/// import 'firebase_options.dart';
+/// // ...
+/// await Firebase.initializeApp(
+///   options: DefaultFirebaseOptions.currentPlatform,
+/// );
+/// ```
+class DefaultFirebaseOptions {
+  static FirebaseOptions get currentPlatform {
+    if (kIsWeb) {
+      throw UnsupportedError(
+        'DefaultFirebaseOptions have not been configured for web - '
+        'you can reconfigure this by running the FlutterFire CLI again.',
+      );
+    }
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return android;
+      case TargetPlatform.iOS:
+        return ios;
+      case TargetPlatform.macOS:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for macos - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.windows:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for windows - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      default:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions are not supported for this platform.',
+        );
+    }
+  }
+
+  static FirebaseOptions android = FirebaseOptions(
+    apiKey: Platform.environment['FIREBASE_OPTIONS_ANDROID_API_KEY']!,
+    appId: Platform.environment['FIREBASE_OPTIONS_ANDROID_APP_ID']!,
+    messagingSenderId: '37031961174',
+    projectId: 'go-egypt-9e91b',
+    storageBucket: 'go-egypt-9e91b.firebasestorage.app',
+  );
+
+  static FirebaseOptions ios = FirebaseOptions(
+    apiKey: Platform.environment['FIREBASE_OPTIONS_IOS_API_KEY']!,
+    appId: Platform.environment['FIREBASE_OPTIONS_IOS_APP_ID']!,
+    messagingSenderId: '37031961174',
+    projectId: 'go-egypt-9e91b',
+    storageBucket: 'go-egypt-9e91b.firebasestorage.app',
+    iosBundleId: 'com.example.goEgypt',
+  );
+}
